refactor(codegen): separate EVM IR from assembly program
